Ported features from Modern Emerald + Working challenges + Fixes (#16)
* Fixed Gen 8 Synchronize

* Ported features from Modern Emerald 2.3.2

* Ported features from Modern Emerald 2.3.3

* Ported features from Modern Emerald 2.3.4

* Ported features from modern Emerald 2.4

* Ported features from Modern Emerald 3.0

* Ported features from Modern Emerald 3.1b

* Ported features from Modern Emerald 3.2

* Fixed and improved options menu

* Edited "Even Faster Joy" option for consistency with HnS

* Edited the "Modern Bg" for consistency with HnS

* Edited, improved and tested all the challenges from Modern Emerald

* Disabled unobtainable Pokémon from the Pokédex order list (Jirachi, Regis, Arceus)

---------

+++ src/battle_tower.c
@@ -319,7 +319,7 @@ const u8 gTowerFemaleFacilityClasses[27] =
     FACILITY_CLASS_LEAF
 };
 
-const u8 gTowerMaleTrainerGfxIds[43] =
+const u16 gTowerMaleTrainerGfxIds[43] =
 {
     OBJ_EVENT_GFX_HIKER,
     OBJ_EVENT_GFX_BURGLAR,
@@ -366,7 +366,7 @@ const u8 gTowerMaleTrainerGfxIds[43] =
     OBJ_EVENT_GFX_ARCHER
 };
 
-const u8 gTowerFemaleTrainerGfxIds[27] =
+const u16 gTowerFemaleTrainerGfxIds[27] =
 {
     OBJ_EVENT_GFX_WOMAN_2,
     OBJ_EVENT_GFX_ENGINEER,
@@ -993,6 +993,10 @@ static const u16 sFrontierTrainerIdRanges[][2] =
     {FRONTIER_TRAINER_ZACHERY, FRONTIER_TRAINER_LAMAR},  // 160 - 199
     {FRONTIER_TRAINER_HANK,    FRONTIER_TRAINER_TESS},   // 180 - 219
     {FRONTIER_TRAINER_JAXON,   FRONTIER_TRAINER_GRETEL}, // 200 - 299
+    //{FRONTIER_TRAINER_JAXON,   FRONTIER_TRAINER_CHAMPION_STEVEN}, // 200 - 319
+                                                                    // Should use this string instead, but new trainers added to the BF aren't
+                                                                    // correctly done. For example, Gym Leader Tate is now Champion Blue, but text hasn't
+                                                                    // been edited, neither trainer parties. So, for now, new trainers do not show up in BF.
 };
 
 static const u16 sFrontierTrainerIdRangesHard[][2] =
@@ -1005,6 +1009,11 @@ static const u16 sFrontierTrainerIdRangesHard[][2] =
     {FRONTIER_TRAINER_JAXON,   FRONTIER_TRAINER_TESS},   // 200 - 219
     {FRONTIER_TRAINER_LEON,    FRONTIER_TRAINER_RAUL},   // 220 - 239
     {FRONTIER_TRAINER_JAXON,   FRONTIER_TRAINER_GRETEL}, // 200 - 299
+    //    {FRONTIER_TRAINER_KAY,     FRONTIER_TRAINER_CHAMPION_STEVEN}, // 240 - 319
+    //    {FRONTIER_TRAINER_HUGO,    FRONTIER_TRAINER_CHAMPION_STEVEN}, // 280 - 319
+                                                                        // Should use these 2 strings instead, but new trainers added to the BF aren't
+                                                                        // correctly done. For example, Gym Leader Tate is now Champion Blue, but text hasn't
+                                                                        // been edited, neither trainer parties. So, for now, new trainers do not show up in BF.
 };
 
 // Unknown, unused data
@@ -1293,7 +1302,7 @@ void SetBattleFacilityTrainerGfxId(u16 trainerId, u8 tempVarId)
 {
     u32 i;
     u8 facilityClass;
-    u8 trainerObjectGfxId;
+    u16 trainerObjectGfxId;
 
     SetFacilityPtrsGetLevel();
     if (trainerId == TRAINER_EREADER)
